requirements.txt
django
asgiref
django-widget-tweaks
Pillow
pytz
sqlparse
django-ckeditor
django-cors-headers
requests
cryptography